requirements.txt
django==2.2
drf-haystack 
elasticsearch
djangorestframework